+++ CHANGELOG.md
@@ -56,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change `__dataclass_params__` to `_dataclass_params`
 - Make component reload after caching in apply
 - Fix a minor bug in schedule_jobs
+- Fix vector index cleanup
 
 ## [0.3.0](https://github.com/superduper-io/superduper/compare/0.3.0...0.2.0])    (2024-Jun-21)